Fix test

+++ gpMgmt/test/behave/mgmt_utils/gpcheckcat.feature
@@ -632,8 +632,9 @@ Feature: gpcheckcat tests
         And gpcheckcat should not print "SUMMARY REPORT: FAILED" to stdout
         And gpcheckcat should not print "has a dependency issue on oid" to stdout
         And gpcheckcat should print "Found no catalog issue" to stdout
+        And the user runs "psql -d check_dependency_error -c "DROP OWNED BY foo; DROP ROLE foo""
+        Then psql should return a return code of 0
         And the user runs "dropdb check_dependency_error"
-        And the user runs "psql -c "DROP ROLE foo""
 
     Scenario: gpcheckcat should not report inconsistency because of scram-sha-256 passwords
         Given database "check_scram_passwords" is dropped and recreated
@@ -644,8 +645,9 @@ Feature: gpcheckcat tests
         And gpcheckcat should not print "SUMMARY REPORT: FAILED" to stdout
         And gpcheckcat should not print "inconsistent_pg_authid" to stdout
         And gpcheckcat should print "Found no catalog issue" to stdout
+        And the user runs "psql -d check_scram_passwords -c "DROP ROLE foo""
+        Then psql should return a return code of 0
         And the user runs "dropdb check_scram_passwords"
-        And the user runs "psql -c "DROP ROLE foo""
 
 
 ########################### @concourse_cluster tests ###########################
